+++ finished/steptwo.mjs
@@ -8,8 +8,8 @@ const invertimage = (filename) => {
     console.log("File format not supported");
     return 1;
   }
-  const width = data.readUint32LE(18);
-  const height = data.readUint32LE(22);
+  const width = data.readUint16LE(18);
+  const height = data.readUint16LE(22);
   const imageData = data.slice(52, data.length);
   console.log("Width is " + width);
   console.log("Height is " + height);
